edit parse type

+++ snack-4/script.js
@@ -5,10 +5,10 @@
 let array = [];
 
 for (let i = 1; i <= 6; i++) {
-  let number = parseInt(prompt(`Inserisci numero ${i} di 6:`));
+  let number = parseFloat(prompt(`Inserisci numero ${i} di 6:`));
 
   if (!isNaN(number)) {
-    number = parseInt(number);
+    number = parseFloat(number);
     if (number % 2 !== 0) {
       array.push(number);
     }
